+++ contracts/exchange/test/utils/index.ts
@@ -1,5 +1,4 @@
 export * from './exchange_wrapper';
-export * from './fill_order_combinatorial_utils';
 export * from './match_order_tester';
 export * from './exchange_data_encoder';
 export * from './types';
