+++ contracts/test-utils/src/index.ts
@@ -29,30 +29,18 @@ export { bytes32Values, testCombinatoriallyWithReferenceFuncAsync, uint256Values
 export { TransactionFactory } from './transaction_factory';
 export { testWithReferenceFuncAsync } from './test_with_reference';
 export {
-    AllowanceAmountScenario,
-    AssetDataScenario,
-    BalanceAmountScenario,
     ContractName,
     ERC20BalancesByOwner,
     ERC1155FungibleHoldingsByOwner,
     ERC1155HoldingsByOwner,
     ERC1155NonFungibleHoldingsByOwner,
     ERC721TokenIdsByOwner,
-    ExpirationTimeSecondsScenario,
-    FeeAssetDataScenario,
-    FeeRecipientAddressScenario,
     FillResults,
-    FillScenario,
     MarketBuyOrders,
     MarketSellOrders,
-    OrderAssetAmountScenario,
     OrderInfo,
-    OrderScenario,
     OrderStatus,
-    TakerAssetFillAmountScenario,
-    TakerScenario,
     Token,
     TokenBalancesByOwner,
-    TraderStateScenario,
     TransactionDataParams,
 } from './types';
